requirements.txt
streamlit
firebase-admin
google-generativeai
spotipy
fpdf
python-dotenv
